component/table: Update showcols to use keys
This updates the table component to use keys to keep track of which
columns should be visible instead of column indices. This will be
necessary for keeping track of which columns were selected after a
component has been unmounted and re-mounted again later. (e.g. for
persistence or saving of user preferences)