Create settings package
This creates a package to be known as `@woocommerce/settings`. This reads the wcSettings global and exports known shared variables, setting defaults if those values are missing.

Components will use this package to get settings from the server, instead of looking up values themselves.

WC-Admin specific variables such as endpoint data is not included in this package. It's mostly for components and common values.