Fix "Back to..." link disappearing after editing a term.
After editing a taxonomy term (e.g. a category or a tag), a success message
with a "Back to <taxonomy>es" link appears, but it disappears after
a couple of seconds.

The problem is that `Loader.php` defines an `admin_notices` that wraps
all notices in a hidden div. This div includes another one with class
`wp-header-end`. There's a piece of code in `common.js` in WP core that
 moves all notices inside that inner div, so as soon as that js executes,
the notice is hidden.

Notice divs having the `inline` class are not moved, therefore the fix
consists of a one-line JavaScript that adds that class to the notice div
before `common.js` executes.