Fix a React warning in test that the key prop wasn't assigned in List
An alternative approach I considered was giving all the data used
in tests a key prop because that also would have solved it, but
it made me wonder if all these types of data even have key props
on them. Without any types to validate something like that, I felt
the safer choice was just to use the array index as the key for the
top level item instead.