Add the setup tab to <ActivityPanel>
It was simpler to maintain a separate list of visible tabs for
situations where we display the setup progress tab.

It was a nice to have to display the task list, but I had some
issues with that, which will require some refactoring of those
components and I've decided to avoid doing that here.

For now we support the most basic functionality which is routing
to the home page where the setup task list exists.