Update wp-prettier to latest beta release to avoid a crash in Prettier
In the previous release Prettier did not work in VSCode (and I
suspect would fail in other editors too) with an error:

"Could not find module ./parser-babylon". Updating to the latest
version fixes it in my environment.

Testing details:

* Ensure you have done a fresh npm install
* Edit a JS file (ideally in VSCode with a prettier plugin if you can)
* Save it (if your prettier plugin has format on save)

You should not see errors from the plugin (in VSCode the plugin reports
status in the bottom bar) and it should format the file correctly.