Remove the chevron icons from the “Finish setup” task list
This aligns with the update of the Store Management card in GH-4790 by removing the chevron icon from the item rows in this task list.

A small refactor was made to remove the conditional operator when assigning `task.after`. This is because we no longer need to render anything if the task is not dismissible.