Add check for remote inbox notifications opt-out (#5162)
The marketplace suggestions opt-out is now used to determine if remote inbox notifications should be fetched. This corresponds to the
`woocommerce_show_marketplace_suggestions` WP option.

Additionally, a new protected method is added to the class. This provide a series of checks for whether remote notifications are enabled on the site.