Remove store registration checks that are no longer needed. (#5170)
Fixes #4443

Some investigation found that #4443 is no longer a reproducible issue. Why this issue no longer exists is not clear, but it makes sense now to disable the store registration checks as they are only called once. This PR just removes the checks but does not change any other behaviour.