Removing breadcrumbs from wc-admin header, replacing with simple titl… (#5232)
Removing breadcrumbs from wc-admin header, replacing with simple title (#3963)

This is an interim step towards the integration of wc-navigation, where a back button will be provided to restore the critical functionality of the breadcrumbs. Note that the `breadcrumbs` data was preserved since it's still being utilized do construct the document title (in semi-breadcrumb-fashion) on both the backend and front-end.