Attempt to check for a theme specific element when completing Theme step. (#5367)
The theme selection page e2e test is intermittently failing and it
is not clear why. It could be that there is a failure going from
the last setup step to this one because the selectors for this
step are ambiguous and could refer to anything.

This specifically waits for an element that is unique to the
theme selection screen in the hope of removing ambiguity.