Fix missing styles on buttons
The <Button> component styles from @wordpress/components were missing color styles. This is due to a change in @wordpress/base-styles where a different technique for providing the default button colors was implemented in https://github.com/WordPress/gutenberg/pull/24408.

As a temporary measure, this restores the previous code by copying it here directly.